Reformatting everything into FlightDatabaseUtils
